@@ -1,17 +1,17 @@
 name: git_hooks
 description: Using dart to prevents bad commit or push (git hooks, pre-commit/precommit, pre-push/prepush, post-merge/postmerge and all that stuff...).It is similar to husky.
-version: 0.1.4
+version: 0.2.0
 homepage: https://github.com/xuzhongpeng/git_hooks
 
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  path: ^1.7.0
-  yaml: ^2.2.1
+  path: ^1.8.0
+  yaml: ^3.1.0
 
 dev_dependencies:
-  pedantic: ^1.9.0 
+  pedantic: ^1.11.0
 executables:
   git_hooks:
