@@ -12,12 +12,12 @@ class Utils {
   }
 
   /// get path of git_hooks library
-  static String getOwnPath() {
+  static String? getOwnPath() {
     var pacPath = path.fromUri(path.current + '/.packages');
     var pac = File(pacPath);
     var a = pac.readAsStringSync();
     var b = a.split('\n');
-    String resPath;
+    late String resPath;
     b.forEach((v) {
       if (v.startsWith('git_hooks:')) {
         var index = v.indexOf(':');
