@@ -11,7 +11,7 @@ String _rootDir = Directory.current.path;
 /// install hooks
 class CreateHooks {
   /// Create files to `.git/hooks` and [targetPath]
-  static Future<bool> copyFile({String targetPath}) async {
+  static Future<bool> copyFile({String? targetPath}) async {
     if (targetPath == null) {
       targetPath = '/git_hooks.dart';
     } else {
@@ -62,7 +62,7 @@ class CreateHooks {
       var hookTemplate = hookFile.readAsStringSync();
       var match =
           RegExp(r'dart\s(\S+)\s\$hookName').allMatches(hookTemplate).first;
-      commandPath = match.group(1);
+      commandPath = match.group(1)!;
       return false;
     });
     return commandPath;
